session: Only emit verification-complete on reauth or after session is opened
We're currently emitting verification-complete signal as soon as we
got a positive authentication callback from the worker, however at this
point we've not opened the session yet, and if that fails we'll end up
in a scenario where gdm assumed that we're logging in, and so its
greeter is shut down and never restarted on its VT until we manually
switch back and forth from it. Meanwhile the session opening failure is
never exposed to the user.

So, do not emit verification-completed on logins, until we've opened
the session.
However, re-order the session-opened callback events so that we keep
notifying verification-complete before than session-opened.

This basically reverts commit c4292870 (again).

Fixes: https://gitlab.gnome.org/GNOME/gdm/-/issues/712


(cherry picked from commit f65c681a469a2675d96012a77a563e50369b9e54)