daemon: Consolidate session-type and supported-session-types list
There's currently a bug in computing the session-type to use.

The `i > 0` check means wayland will overwrite x11 in the
transient session type list.

Morever, the separation between "session-type" and
"supported-session-types" is a little redundant. Since
supported-session-types is a sorted list, the first item should
always be the same as "session-type".

This commit addresses the bug and the redundant logic, by computing
the supported session types early in the function and indexing into
it to get the session-type.

A future cleanup could probably get rid of session-type entirely.

https://gitlab.gnome.org/GNOME/gdm/-/merge_requests/153