pam: Drop gdm-pin service
gdm-pin was an experimental feature that was going to get added to
gnome-shell many years ago. It never happened and these days it
would probably be implemented a little different anyway.

(It would probably use a gdm pam extension)

There's no point keeping this service file around that we aren't
using, so this commit drops it.

Closes: https://gitlab.gnome.org/GNOME/gdm/-/issues/731