daemon: Infer session type from desktop file if user has no saved session type
The accountsservice user cache file can specify a session type
associated with the saved session.  This is optional though. If one
isn't specified GDM needs to figure out the session type based on the
list of preferred session types for the system and the session file
itself.

It was failing to do the latter, though.  This commit fixes that.