gdm.rules: Keep wayland enabled for simple-framebuffer DRM drivers
Not all DRM drivers are disabled when the nomodeset kernel cmdline option
is used. For example the simpledrm driver that use the system framebuffer
set-up by the bootloader, provides a modesetting interface.

Exclude the DRM drivers that match against the "simple-framebuffer" device
and only disable wayland for platform DRM drivers.

This allows to start a wayland session when nomodeset is used to disable a
platform DRM driver by using the simpledrm driver instead of legacy fbdev
drivers such as efifb, that does not support modsetting and could only be
used with an Xorg session.