data: Don't race with vendor nvidia driver at startup
The vendor nvidia driver has two modules loaded
at startup.

It's not uncommon for the second module to still
be loading when GDM starts.

Our udev rule relies on the second module to set
up GDM's boot configuration.

This commit adds some synchronization to stall
GDM a bit if the vendor driver is still coming
up.

https://gitlab.gnome.org/GNOME/gdm/-/issues/763