common: Add API to reload settings from disk
Ideally we would reread /run/gdm/custom.conf after we've decided
graphics setup is complete.  This is because the file may not
get written out by udev until after GDM is already started and waiting.

As a first step to get there, this commit adds an API for rereading
the file, and changes the SIGHUP handler to use it (instead of
the complete teardown and reinitialization it was doing before).