local-display-factory: Stall startup until main graphics card is ready
At the moment, GDM waits until systemd says the system supports
graphics (via the CanGraphical logind property).

Unfortunately, this property isn't really what we need, since it flips
to true when *any* graphics are available, not when the main graphics
for the system are ready.

This is a problem on hybrid graphics systems, if one card is slower to
load than another. In particular, the vendor nvidia driver can be slow
to load because it has multiple kernel modules it loads in series.

Indeed on fast systems, that use the vendor nvidia driver, it's not
unusual for boot to get to a point where all of userspace up to and
including GDM is executed before the graphics are ready to go.

This commit tries to mitigate the situation by adding an additional,
check aside from CanGraphical to test if the system is ready.

This check waits for the graphics card associated with boot to be fully
up and running before proceeding to start a login screen.

Closes: https://gitlab.gnome.org/GNOME/gdm/-/issues/763