meson: set -D_GNU_SOURCE for updwtmpx
Without setting GNU_SOURCE, we end up getting:
```
../gdm-42.0/daemon/gdm-session-record.c:200:9: error: implicit declaration of function ‘updwtmpx’; did you mean ‘updwtmp’? [-Werror=implicit-function-declaration]
updwtmpx (GDM_NEW_SESSION_RECORDS_FILE, &session_record);
```

This ended up exposing a bug in updwtmp(3) (which is now fixed
thanks to the man-pages maintainers!) as it didn't mention that updwtmpx
is a GNU extension (and hence needs GNU_SOURCE in order to be available).

Alternatively, we could just #define _GNU_SOURCE in gdm-session-record.c
for updwtmpx.

Bug: https://bugzilla.kernel.org/show_bug.cgi?id=216168