Introduce Remote Display Factory
This commit adds a new class used to create remote displays.

It exposes a RemoteDisplayFactory interface over the bus to create remote displays
with a CreateRemoteDisplay method. This method sets the received remote_id to
on the created RemoteDisplay.

When it creates a remote display, it stores it in the display store and
prepares it (which makes the remote display create a headless login screen)