settings: Add RemoteLoginEnable option
As a first step toward getting remote login support in GDM, this
commit adds a new RemoteLoginEnable option.

This is a big switch to just turn the feature on or off. The
switch is on by default, because it doesn't actually
enable remote login by itself. It still requires gnome-remote-desktop
(or whatever) to be running as a system service to manage the remote
login requests.