Introduce Remote Display
New class that is used to define wayland headless remote displays.

These displays expose a dbus interface RemoteDisplay with RemoteId property
which represents the id of the RemoteClient from grd daemon-system.
This means this display will be used by that RemoteClient.

Bind the session_id property from the display parent and expose it on
the RemoteDisplay interface too.