manager: Fix btmp record accounting
Before a user logs in they don't have a display.

btmp records currently need a display though, and they
get written when the user can't log in.

Furthermore, the display from X11 point of view is
somewhat archaic. We use wayland by default now.

In lieu of a display, this commit gives the btmp record
the seat id instead.