pam-extensions: Separate ChoiceList extension to its own header
We're going to be adding another extension or two in the near future,
so having everything in one header file isn't great for structure's
sake.

This commit moves the common code to a common header, the extension
to its own header, and makes gdm-pam-extensions.h just include
the other headers.