daemon: Specify DCONF_PROFILE for initial-setup
Since commit edc1815175, the initial-setup session no longer uses
the locked down "gdm" profile, but the regular user profile.

On the one hand this seems justified, given that the initial-setup
session is different enough from the login screen, however it is
also different from a regular user session:

 - screen locking isn't useful for a system user
 - during setup, log out / user switching are pointless at best
 - accessibility features should be readily available in case
   the user needs them

For this reason, return to specify DCONF_PROFILE for the initial-setup
user, but use a distince profile from "gdm" to allow initial-setup
to provide its own presets.