manager: Create remote displays for user sessions started from remote login screens
If a login screen is running on a remote display, the session that gets
started from that login screen should have a remote display associated
with it as well, instead of a local display object.

Furthermore, the two remote displays should have a comment remote_id so
there's a clear trail from one to the other.

This commit does that.