gdm-client: Add user verifier PAM extensions getter
We have a setter to enable the user verifier PAM extensions, but it's
not possible to get back this parameter.

So add a getter for it. We're still returning the value that the library
user defined, more than the extensions that are actually enabled as that
may lead to a racy result, given that it depends on interaction with
GDM session.