Add support for generic GDM CustomJSON binary protocol
GDM allows passing data through PAM binary conversations, however it
only supports an extension protocol for choice list.

Due to the GDM architecture, creating a new custom protocol is quite
verbose and it implies changing multiple layers, at the same time it is
impossible for modules and GNOME Shell GDM/login extensions to easily
define specific protocols to perform custom actions, and they currently
need to rely on DBus or other ways to communicate without an easy way to
setup a secure communication channel.

At the same time, extending a binary protocol is hard, once one has
been defined.

To make this easier to manage both by modules and by gdm clients, add
a simple JSON-based custom protocol that it's easy to extend and
that moves the protocol definition at higher level than gdm.

This moves the control of custom JSON protocols out from GDM, leaving
to the modules and to gdm clients (the shell) the control over them.

GDM is still in charge of doing some validation check of the JSON
provided by the two parties, but it won't be in charge of any further
protocol operation.