manager: Use full verification on incompatible login session
Two sessions are incompatible if any is remote or don't share the same
seat_id.

From a login session, when the logged in user session was previously
started and is incompatible with this login session, don't do the
reauthentication process and do the full verification instead. This will
be used in next commits to shutdown the previously started session and
start a new session.