Let temp reauth switch back to user session
The temporary reauth channel was used to unlock sessions not directly
managed by GDM (i.e. those running in VNC, or similar). It was not
accessible from the login screen. Thus, there was no need to switch the
display back over to the user's session.

Now that temporary reauth channel can be used from the login screen for
users backed by systemd-homed, this commit adds the functionality to
switch back to the existing user session