gdm-session-worker: Handle PAM_MODULE_UNKNOWN as service unavailable
In case a service file is there but a library it's loading is missing,
still return the module unavailable error that has benefit in stopping
the shell for complaining about the service.

In debian-based systems pam services may be sticking to a location so
even though if a module it uses may have been removed.

So handle this case gracefully