manager: Update remote-id on reauth from RemoteDisplay
Now two headless sessions are compatible.
On reauthentication, the login display will get the property
of its reauth_caller_pid.
Now when a session is reauthenticated, on "Reauthenticated" signal it
is given the pid_of_caller property. This way it can be known from which
login display it has been reauthenticated, when multiple login displays
attempted to reauth the same user.
On succesfull reauthentication, get the login display using that
property. If that display is remote, set the "remote-id" of the login
display to the user display, this makes the g-r-d daemon-system, handover
an RDP client to that reauthenticated session.