Fix returning to session from temp reauth channel
The previous version of this code was simply wrong. It would pass the
session we want to return to into switch_to_compatible_user_session, but
it would pass TRUE to fail_if_already_switched. This would exclude the
session we want to activate from the list of sessions we can activate.
This leaves the user stranded on a frozen GDM when reauthenticating
during a secure lock