remote-display-factory: Block suspend when any remote display exists
The remote display factory has a counter of how many remote displays
are. With that counter when there's at least one, block suspend, when
there's none unblock suspend.

This is necessary because other sessions might want to suspend without
being aware that there might be remote sessions being used. On suspend
the RDP client would disconnect and that makes the remote session die.
In the future the sessions would not die on disconnecting, but still is
undesirable the RDP client disconnecting.