udev: Remove rules for disabling Wayland if nomodeset
This rule is wrong. The way it's written it will Wayland iff:
1) nomodeset is passed
2) some DRM device registers
3) the DRM device is not SimpleDRM

The logic is wrong because:
* on systems without SimpleDRM rule 2 would fail
* on nvidia, which ignores `nomodeset`, rule 3 would match

So it is not disabling Wayland when it should, and it is when it shouldn't.

We don't need to disable Wayland now, we can let GDM try and if it does
not find any suitable DRM device it will automatically fall back to Xorg.