gdm-wayland-session: Register display optionally
Before we'd always register the display as soon as the session was
started. However, this is too early for Wayland! The compositor might
not have completely initialized its Wayland connection yet, so the
display isn't really open.

First, we added a delay to the automatic registration. Now we give the
compositor three seconds to start itself before we declare the display
functional.

Next, we give the compositor an option to opt out of this automatic
registration. Instead, the compositor will be responsible for
registering itself when it's up and running.