manager: Combine register display with register session
Before we'd always register the display as soon as the session was
started. However, this is too early for Wayland! The compositor might
not have completely initialized its Wayland connection yet, so the
display isn't really open.

The "register display" should happen when the compositor is up and
running. This occurs when it calls RegisterSession, so it makes sense
to combine both.

This helps on terminating plymouthd when the new session has been
started, and also kill background greeters.

Part-of: <https://gitlab.gnome.org/GNOME/gdm/-/merge_requests/285>