Introduce gdm-new-session
This daemon/tool allows starting a user session either locally or headlessly.

It communicates with the GDM daemon to start/stop the sessions using
the dbus ifaces RemoteDisplayFactory and LocalDisplayFactory.

It watches the GDM Displays dbus object manager to:
  1. Check when a new session is created to ensure it has
     successfully started the session for the user.
  2. Check when the session is finalized to terminate itself.

When it finds out it has successfully started the session, it stores the
session_id. This allows terminating the session in the case this is
terminated.

Install also a service that allows starting a headless user session:
gnome-headless-session@.service

This provides the same functionality than TigerVNC vncserver@.service:
  vncserver@.service allows to start a headless Xvnc session for a user.
  There's a conf file that maps a display number and how that display
  session should be, its username...
  It is started as a system service. e.g. vncserver@:1.service

In the case of gnome-headless-session@.service, there's no conf file.
It is a system service that starts a headless gnome desktop session
for the user passed as the argument. e.g. gnome-headless-session@joan.service

In order to get remote access, it is needed to:
  1. Configure gnome-remote-desktop for the user with grdctl --headless ...
  2. Enable gnome-remote-desktop-headless.service.

Now the polkit rule to use the dbus methods for create/destroy user
displays is updated and only authorizes processes run from the unit
service gnome-headless-session@.service and for the user passed as the.
argument.

Part-of: <https://gitlab.gnome.org/GNOME/gdm/-/merge_requests/288>