remote-display: Add SetRemoteId method
Now accessing these created headless user sessions has two different ways:
  1. Directly connecting to the gnome-remote-desktop daemon headless
     which is connected to that user session, as explained in the
     previous commit.
  2. Through the remote login. i.e. connecting to the gnome-remote-desktop
     system daemon that provides a headless login screen and after
     authenticating, get that headless user session.

To make the option 2 work, the system daemon needs to know the remote_id
of the headless display to be able to do the handover from the login
headless display to the user headless display.

This allows the system daemon generate a valid remote_id and assign it to
this new headless display.

Part-of: <https://gitlab.gnome.org/GNOME/gdm/-/merge_requests/288>