manager: Also move g-i-s home dir when chowning
Before, there was one gnome-initial-setup user, so both GDM and the
final user session could find the g-i-s homedir. GDM would chown the
files on behalf of the session, and the session's
gnome-initial-setup-copy-worker.service would copy the files over.

Now that we're using dynamic users, not only is the gnome-initial-setup
user's username unpredictable, but it's also quite likely that the user
no longer exists by the time g-i-s-copy-worker runs.

With this commit, GDM not only chowns the files, but also moves them
into a stable location (/run/gdm/gnome-initial-setup). This allows
g-i-s-copy-worker to find them