launch-environment: Add per-seat settings persist
This is a bit of a hack to deal with the regression in GNOME 49, where
we no longer keep any settings because of the GDM dynamic users. For
now, we will only ever run one dynamic user per seat, so we can get away
with this.

Of course this isn't a proper solution for this issue. That would be
https://gitlab.gnome.org/GNOME/gdm/-/issues/993. But we can't do that
for the 49 cycle, so for now this workaround will manage

We also migrate the existing settings from /var/lib/gdm to the seat0
persist location. This means that existing settings that users have set
(most importantly: accessibility settings such as screen-reader-enabled)
will continue functioning through the transition.

Fixes https://gitlab.gnome.org/GNOME/gdm/-/issues/1005