launch-environment: Migrate via copy, not move
This makes it possible to downgrade GDM to a version that predates 49.
This will roll back changes to settings, but presumably if someone is
rolling back they're OK rolling back to their known-working
configuration.

This is important because otherwise someone with a screen reader may
upgrade GDM, run into issues, downgrade, and end up without a screen
reader and thus no ability to use the machine