display: Drop x11-display-name
First of all, this property is never set so it is always NULL. Thus:
this property is dead code.

The `GetX11DisplayName` D-Bus API isn't used by anything. I searched on
Debian Code Search and GitHub and found no users. Looks like it was used
historically by a GDM component called "gdm-slave" but this no longer
exists in the modern GDM architecture