Finalize removal of GDM user
There is no longer a GDM user. In GNOME 49 we switched to using dynamic
users, but we couldn't completely remove the old GDM user because it was
used by the remaining X11 bits:

- The legacy X11 server ran the /etc/gdm/Init hook scripts as the
GDM_USERNAME user

- All the X11 hook scripts were provided an Xauthority that allowed the
GDM_USERNAME to talk to the X11 server (this only worked with the legacy
X11 server, since the per-user X11 server wouldn't be running yet)

Now that all of this is gone, we can completely drop all references to
the user!