common: Use Xwayback if available
Wayback[1] is an X11 compatibility layer that lets you run full
traditional X11 desktop environments on top of Wayland. It consists of a
special stripped-down Wayland compositor that's purpose built for only
running Xwayland, which then acts like a normal X server for the rest of
the session.

Wayback is important because soon hardware will lack native support for
Xorg. For instance: some GPUs like those in Apple Silicon Macs have
drivers with no X11 support whatsoever. Wayback allows desktops that
still use X11 to leverage the platform's strong Wayland support to
function, despite the lack of native Xorg support.

This will also become increasingly important as Xorg drifts into
irrelevance and becomes unmaintained. Xwayland will remain maintained
for the foreseeable future, and Wayback allows desktops that require X11
to leverage this ongoing maintenance.

Wayback is still a bit experimental, but also it won't be installed by
users and/or distros unless it's necessary and/or ready. So we can use
it if it's present, and fall back to Xorg if it's not.

[1]: https://wayback.freedesktop.org/