launch-environment: Let systemd pickup greeter environment variables
When the PAM session is opened, the `systemd --user` daemon starts and will
activate all units part of default.target.
By the time we pass our environment to the gnome-session leader process, which
will export all the environment to systemd, the services part of default.target
will already be running with an incomplete environment.

Normally this is not an issue as GDM will only set variables relevat to a
graphical session. But in case of the greeter session we need to be able to set
generic environment variables such as XDG_STATE_HOME and XDG_CONFIG_HOME, such
that all services, even those started with default.target, can make use of the
persistent storage.

Notably, this fixes the case where pipewire/wireplumber are started with
default.target and wouldn't otherwise be able to restore output and volume
configuration when starting the greeter again.