gdm-x-session: Remove the option parsing
Drops the three options we pass into gdm-x-session:

`--register-session` was controlled by `X-GDM-SessionRegisters`, which
is used only by GNOME (and, seemingly incorrectly, by the occasional
window manager on GitHub). Since there's no more GNOME on X11, this will
always be false

`--run-script` was controlled by X-GDM-BypassXsession", which is set by
nothing. I did a code search and only found some usages in ancient
Solaris fallback-to-xterm scripts, and in historical forks of nixpkgs
(modern nixpkgs doesn't use it). So this will always be true

`--allow-remote-connections` is controlled by display_is_local and a
setting. If you do a truth table you'll discover that the
display_is_local variable has exactly 0 influence on the outcome of the
option: the option is entirely controlled by the setting. So I just made
gdm-x-session check the setting on its own

This allowed me to drop the option parser from gdm-x-session