display: handle XFree86_VT property type of "None"
When a property name is known, but the property is not set on the given
window, the X server will respond with a property type of "None" and no
value.  This fixes the code to handle that, instead of trying to read
the non-existent property value from past the end of the allocated reply.

Fixes: ca9006193 ("display: port GdmDisplay to xcb")
Closes: #1038
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>