libgdm: Track proxies connections and unset them on disconnect
If one of the gdm_client_get_* functions to get dbus proxies is used
after the proxy connection is closed, we were still returning it, but
in an invalid state.

This can happen when gdm fails to start the session application for
whatever reason, in fact in such case gnome-shell is not restarted and
so it preserves the same gdm client instance, but meanwhile the proxies
connections have been closed.

Now, in such state, the gdm library would still return the old cached
invalid proxies, since we do not really check for the connection
validity.

To prevent this to happen, track the connection state of the cached
proxies and nullify the proxies instances on disconnection.

Co-authored-by: Marco Trevisan (Treviño) <mail@3v1n0.net>
Part-of: <https://gitlab.gnome.org/GNOME/gdm/-/merge_requests/345>