libgdm: Track proxies connections and unset them on disconnect
There can be various situations where the proxies connection is closed,
e.g. when gdm fails starting the session and the login session isn't
restarted, meaning it has the same gdm client with the same proxies
(now invalid).

In those cases, when gdm_client_get* funcs are called, invalid proxies
would be returned.

To fix this, track the proxies connections and unset them when the
connection is closed.

Co-authored-by: Marco Trevisan (Treviño) <mail@3v1n0.net>